[#34] refactor: move all jest unit-tests to their own folder
